handle replicated databases
